refactor: Thread Context - Add ThreadComposerState/Store
Updated the import paths in several files to import from ComposerStore instead of ComposerTypes for consistency and clarity.
